requirements.txt
psycopg2-binary==2.9.6
python-environ==0.4.54
requests==2.28.2
SQLAlchemy==2.0.8